### 2022-07-02 #4544 @editorialbot
[PRE REVIEW]: DPFEHM: a differentiable subsurface physics simulator
**Submitting author:** @omalled (<a href="http://orcid.org/0000-0003-0432-3088">Daniel O&#39;Malley</a>)
**Repository:** https://github.com/OrchardLANL/DPFEHM.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
