### 2022-07-03 #4544 @Kevin-Mattheus-Moerman
@omalled thanks for this submission to JOSS. I am an AEiC with JOSS and here to help with the initial steps (and with finding a handling editor). I attempted to fix the compilation error with this PR: https://github.com/OrchardLANL/DPFEHM.jl/pull/8, can you have a look? Once you've implemented the fix you can call `@editorialbot generate pdf` to attempt paper compilation again. Thanks
