### 2022-07-03 #4544 @jedbrown
This looks interesting and going in a direction that I think is very fruitful. Notes/questions from my brief survey:
1. I see there are great examples, but the actual package is only 543 LOC. This usually triggers a scope query.
2. It would be helpful for the paper to describe capabilities in comparison with existing subsurface flow software and any major assumptions in the present model. How difficult would it be to support tensor-valued conductivity or non-orthogonal grids (needs multi-point flux)? How about a different equation of state, perhaps with trainable parameters?
3. I notice a TODO related to handling capillary pressure.
4. Are there any performance claims relative to common solvers for such problems?
5. What is your verification and validation strategy?
