### 2022-07-04 #4544 @omalled
Thanks for taking a look. Here's my perspective on the points you raised:

1. There are a couple of companion packages that were built to support DPFEHM. One is [DifferentiableBackwardEuler](https://github.com/OrchardLANL/DifferentiableBackwardEuler.jl), which does the time-integration of the ODEs in a differentiable way. The other is [NonlinearEquations](https://github.com/OrchardLANL/NonlinearEquations.jl), which does sparse automatic differentiation, and was intended to reduce the number of lines of code developers need to write to build PDE solvers like these. I don't include them in the DPFEHM repo, since they could be useful in their own right. When I give talks about DPFEHM, I have slides showing that the `@NonlinearEquations.equations` macros (e.g., [here](https://github.com/OrchardLANL/DPFEHM.jl/blob/884c05d4417b0fb932cda51f1addc0b48c440792/src/richards.jl#L57)) expand ~20 lines of code describing the equations into ~500 lines of code that compute the residuals and the sparse Jacobians. That's how many lines of code you'd need to write (for each type of physics) doing this the usual way. Overall, this was a lot more than 3 months of work for an individual, which was the criteria I found on the JOSS website.
2. This is a good point -- I'll add more about comparisons and assumptions in the paper. In principal, it shouldn't be hard to do things like multi-point flux. However, I'm not experienced with those methods and my focus has been on the differentiability angle, so I kept the numerical methods vanilla (two-point flux approximation). I think it should be possible to do a trainable equation of state but we haven't played around with that.
3. Yes, something I've learned about subsurface flow solvers is you could always add more physics. :smile:
4. I have done a performance comparison with PFLOTRAN on some flow simulations in discrete fracture networks that a colleague had sitting around (which led to [this paper](https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2021WR031188)). DPFEHM was giving a pretty big speed-up on those problems (~20X, IIRC). However, I attribute that to using more appropriate linear solvers than my colleagues were using with PFLOTRAN (they were using BiCG with an ILU preconditioner, whereas I was using CG with an AMG preconditioner, IIRC). The place where you'd really get a performance advantage is computing the gradient of a loss function involving a subsurface flow simulation. E.g., [this repo](https://github.com/OrchardLANL/DPFEHMWaffle.jl) includes a model that was part of a big project related to the groundwater at LANL. When that project was going on, we were using [FEHM](https://fehm.lanl.gov) (as opposed to DPFEHM) to do the simulations. To keep the cost of computing the gradients manageable, we had to reduce the number of parameters from the natural parameterization (1 permeability per grid node, so O(10^6) parameters) to something much smaller, like O(100) parameters. Even with the reduction in the number of parameters, to compute that gradient in a reasonable time we had to use hundreds of CPUs. With DPFEHM, we could compute that gradient in about the same amount of time on a laptop. We could also compute the gradient of the natural parameterization (with O(10^6) parameters) on a laptop in about the same amount of time. So, that's where the real advantage is. In the jargon of computational science, it's the difference between using the adjoint method to compute the gradient (fast) and using finite differencing to compute the gradient (slow). I'll add some discussion to the paper along these lines.
5. There are a number of [tests](https://github.com/OrchardLANL/DPFEHM.jl/tree/master/test) that we've developed to ensure we don't break things. E.g., comparing to things like the Theis solution and the Thiem solution. I've separately done tests to compare with PFLOTRAN on the discrete fracture networks I mentioned earlier. I've also done more of a validation using the example [here](https://github.com/OrchardLANL/DPFEHMWaffle.jl), which comes from a real-world problem. I don't include these more complex tests in the repo's test set, because they require a lot of supporting data and are too slow to test quickly. Still, they've helped build confidence that the models are doing the right thing, even for complex problems.
