### 2022-07-04 #4544 @omalled
Regarding point 1, I'd also note these earlier packages, which are basically prototypes for DPFEHM and NonlinearEquations: [FiniteVolume](https://github.com/madsjulia/FiniteVolume.jl) and [LinearAdjoints](https://github.com/madsjulia/LinearAdjoints.jl). They explore some of the same ideas, but are not as full-featured. E.g., DPFEHM has a lot more physics than FiniteVolume and LinearAdjoints works for linear equations, whereas NonlinearEquations deals with, as you would guess from the name, nonlinear equations.
