+++ paper.bib
@@ -1,3 +1,10 @@
+@techreport{mercer2020amanzi,
+  title={Amanzi--ATS: Modeling Environmental Systems across Scales [Brief]},
+  author={Mercer-Smith, Janet Anne},
+  number={LA-UR-20-26636},
+  year={2020},
+  institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States)}
+}
 @article{greer2022comparison,
   title={A Comparison of Linear Solvers for Resolving Flow in Three-Dimensional Discrete Fracture Networks},
   author={Greer, SY and Hyman, JD and O’Malley, Daniel},
