### 2022-07-07 #4544 @omalled
Thanks, @WilkAndy, for the quick feedback and I hope you get well soon! Here's my perspective:

1. The name is a little tough. In the early days, I called it a "subsurface flow simulator", then a "subsurface flow and transport simulator" after I added transport. Recently, I went with "subsurface physics simulator" to cover the wave equation simulator. I agree it could be seen as overly broad. Of course, I don't mean to imply it could simulate all possible subsurface physics. But, it does do flow, transport, and the wave equation. I'm not sure what a good umbrella for that is other than "subsurface physics". I am open to suggestions.
2. This is a good observation -- I'll clarify this.
3. Will do.
4. We have an example of that [here](https://github.com/OrchardLANL/DPFEHM.jl/blob/master/examples/pressure_management/ex.jl), which the README.md points to in the "advanced usage" section. Still, I agree it would be better to expand the advanced usage section a bit to demonstrate the capabilities more thoroughly. I will do that.
5. I must have messed up the formatting of the citations. There are supposed to be 3 there. I'll fix that. Thanks for pointing this out.
6. I believe PEST is in a different category. My understanding is that PEST is used to do things like inverse analysis, sensitivity analysis, UQ, etc. given an arbitrary numerical model. By contrast, DPFEHM is the numerical model itself. I could add a discussion of PEST, if desired. Something along the lines of "PEST takes a non-intrusive approach to UQ, etc., which allows it to work with any simulator but forces it to treat the simulator as a black box. DPFEHM lays the groundwork for next-generation UQ tools that utilize the gradient and Jacobian information that DPFEHM efficiently provides." Would something like that help?
