### 2022-07-07 #4544 @WilkAndy
That all sounds good, @omalled .   I'm not sure about (1) !    I won't be able to download, test, play-with your code until next week, sorry, so won't complete my checklist till then.
