### 2022-07-08 #4544 @jedbrown
Ah, I need to issue one command at a time.
