### 2022-07-08 #4560 @omalled
BTW, I added a [GitHub action to run the tests](https://github.com/OrchardLANL/DPFEHM.jl/actions/runs/2632551544).
