### 2022-07-11 OrchardLANL/DPFEHM.jl#11 @WilkAndy
Suggestion: include python as a dependency
You use python/matplotlib for plotting.

Refs https://github.com/openjournals/joss-reviews/issues/4560
