### 2022-07-11 0d8349b
Update README.md
add testing instructions
