### 2022-07-11 ea33300
remove old examples
