+++ examples/groundwater_steadystate.jl
@@ -1,89 +1,0 @@
-import DPFEHM
-import GaussianRandomFields
-import Optim
-import PyPlot
-import Random
-import Zygote
-
-Random.seed!(0)
-
-#set up the grid
-mins = [0, 0, 0]; maxs = [50, 50, 5]#size of the domain, in meters
-ns = [30, 30, 3]#number of nodes on the grid
-coords, neighbors, areasoverlengths, _ = DPFEHM.regulargrid3d(mins, maxs, ns)#build the grid
-
-#set up the boundary conditions
-Qs = zeros(size(coords, 2))
-injectionnode = 1#inject in the lower left corner
-Qs[injectionnode] = 1e-6#m^3/s
-dirichletnodes = Int[size(coords, 2)]#fix the pressure in the upper right corner
-dirichleths = zeros(size(coords, 2))
-dirichleths[size(coords, 2)] = 0.0
-
-#set up the conductivity field
-lambda = 50.0#meters -- correlation length of log-conductivity
-sigma = 1.0#standard deviation of log-conductivity
-mu = -9.0#mean of log conductivity -- ~1e-4 m/s, like clean sand here https://en.wikipedia.org/wiki/Hydraulic_conductivity#/media/File:Groundwater_Freeze_and_Cherry_1979_Table_2-2.png
-cov = GaussianRandomFields.CovarianceFunction(2, GaussianRandomFields.Matern(lambda, 1; σ=sigma))
-
-x_pts = range(mins[1], maxs[1]; length=ns[1])
-y_pts = range(mins[2], maxs[2]; length=ns[2])
-num_eigenvectors = 200
-grf = GaussianRandomFields.GaussianRandomField(cov, GaussianRandomFields.KarhunenLoeve(num_eigenvectors), x_pts, y_pts)
-logKs = zeros(reverse(ns)...)
-logKs2d = mu .+ GaussianRandomFields.sample(grf)'#generate a random realization of the log-conductivity field
-for i = 1:ns[3]#copy the 2d field to each of the 3d layers
-    v = view(logKs, i, :, :)
-    v .= logKs2d
-end
-
-#plot the log-conductivity
-fig, ax = PyPlot.subplots()
-img = ax.imshow(logKs[1, :, :], origin="lower")
-ax.title.set_text("Conductivity Field")
-fig.colorbar(img)
-display(fig)
-println()
-PyPlot.close(fig)
-
-logKs2Ks_neighbors(Ks) = exp.(0.5 * (Ks[map(p->p[1], neighbors)] .+ Ks[map(p->p[2], neighbors)]))
-function solveforh(logKs, dirichleths)
-    @assert length(logKs) == length(Qs)
-    if maximum(logKs) - minimum(logKs) > 25
-        return fill(NaN, length(Qs))#this is needed to prevent the solver from blowing up if the line search takes us somewhere crazy
-    else
-        Ks_neighbors = logKs2Ks_neighbors(logKs)
-        return reshape(DPFEHM.groundwater_steadystate(Ks_neighbors, neighbors, areasoverlengths, dirichletnodes, dirichleths, Qs), reverse(ns)...)
-    end
-end
-
-#now do a forward solve for the head
-print("forward solve time")
-@time h = solveforh(logKs, dirichleths)#solve for the head
-#plot the head at the bottom of the domain
-fig, ax = PyPlot.subplots()
-img = ax.imshow(h[1, :, :], origin="lower")
-ax.title.set_text("Head")
-fig.colorbar(img)
-display(fig)
-println()
-PyPlot.close(fig)
-
-#now compute the gradient of a function involving solveforh
-gradient_node = div(size(coords, 2), 2) + 500
-gradient_node_x = coords[1, gradient_node]
[…]
