+++ examples/groundwater_time_dependent.jl
@@ -1,119 +1,0 @@
-import ChainRulesCore
-import DifferentiableBackwardEuler
-import DPFEHM
-import GaussianRandomFields
-import Optim
-import PyPlot
-import Random
-import Zygote
-
-Random.seed!(0)
-
-#set up the grid
-mins = [0, 0, 0]; maxs = [50, 50, 5]#size of the domain, in meters
-ns = [30, 30, 3]#number of nodes on the grid
-coords, neighbors, areasoverlengths, volumes = DPFEHM.regulargrid3d(mins, maxs, ns)#build the grid
-
-#set up the boundary conditions
-Qs = zeros(size(coords, 2))
-injectionnode = 1#inject in the lower left corner
-Qs[injectionnode] = 1e-6#1e-4#m^3/s
-dirichletnodes = Int[size(coords, 2)]#fix the pressure in the upper right corner
-dirichleths = zeros(size(coords, 2))
-dirichleths[size(coords, 2)] = 0.0
-
-#set up the initial condition
-h0 = zeros(size(coords, 2))
-
-#set up the conductivity field
-lambda = 50.0#meters -- correlation length of log-conductivity
-sigma = 1.0#standard deviation of log-conductivity
-mu = -9.0#mean of log conductivity -- ~1e-4 m/s, like clean sand here https://en.wikipedia.org/wiki/Hydraulic_conductivity#/media/File:Groundwater_Freeze_and_Cherry_1979_Table_2-2.png
-cov = GaussianRandomFields.CovarianceFunction(2, GaussianRandomFields.Matern(lambda, 1; σ=sigma))
-x_pts = range(mins[1], maxs[1]; length=ns[1])
-y_pts = range(mins[2], maxs[2]; length=ns[2])
-num_eigenvectors = 200
-grf = GaussianRandomFields.GaussianRandomField(cov, GaussianRandomFields.KarhunenLoeve(num_eigenvectors), x_pts, y_pts)
-logKs = zeros(reverse(ns)...)
-logKs2d = mu .+ GaussianRandomFields.sample(grf)'#generate a random realization of the log-conductivity field
-for i = 1:ns[3]#copy the 2d field to each of the 3d layers
-    v = view(logKs, i, :, :)
-    v .= logKs2d
-end
-
-#set up the storage
-specificstorage = fill(0.1, size(coords, 2))
-
-#plot the log-conductivity
-fig, ax = PyPlot.subplots()
-img = ax.imshow(logKs[1, :, :], origin="lower")
-ax.title.set_text("Conductivity Field")
-fig.colorbar(img)
-display(fig)
-println()
-PyPlot.close(fig)
-
-logKs2Ks_neighbors(Ks) = exp.(0.5 * (Ks[map(p->p[1], neighbors)] .+ Ks[map(p->p[2], neighbors)]))
-function solveforh(logKs, dirichleths)
-    @assert length(logKs) == length(Qs)
-    Ks_neighbors = logKs2Ks_neighbors(logKs)
-    p = [Ks_neighbors; dirichleths]
-    h_gw = DifferentiableBackwardEuler.steps(h0, f_gw, f_gw_u, f_gw_p, f_gw_t, p, 0.0, 1.0 * 60 * 60 * 24 * 365 * 10; abstol=1e-1, reltol=1e-1)
-    return h_gw
-end
-hflat2h3d(h) = reshape(h, reverse(ns)...)
-
-function unpack(p)
-    @assert length(p) == length(neighbors) + size(coords, 2)
-    Ks = p[1:length(neighbors)]
-    dirichleths = p[length(neighbors) + 1:length(neighbors) + size(coords, 2)]
-    return Ks, dirichleths
-end
-
[…]
