### 2022-07-11 OrchardLANL/DPFEHM.jl#11 @omalled
I'm only using it in the examples, not the DPFEHM module itself. I've seen people in the Julia community viewing `PyPlot.jl` as a brittle package, so having it as a dependency could create headaches. I've had a few of those myself getting matplotlib to play nicely with other python packages. One option might be having a Manifest.toml file in the examples folder that would make it easy to install all the dependences needed to run the examples. Would that help?
