### 2022-07-11 OrchardLANL/DPFEHM.jl#11 @WilkAndy
The `Manifest.toml` might get stale pretty quickly, do you think?  (Unless you have an automated action to create a clean machine, install dependences, run examples on every pull-request.)  i only mentioned python because it's actually a bit of a beast, eg, i'm quite surprised that the python envivonment i'm using actually works with your code, and i didn't have to downgrade/upgrade something.

Having said this, is it possible to create a list of **julia** packages that you need?  Eg, i'm currently getting a conflict between `AlgebraicMultigrid` and `Preconditioners`.  Obviously, you've managed to install non-conflicting versions of these, but i have no idea.
