### 2022-07-11 OrchardLANL/DPFEHM.jl#13 @WilkAndy
ensure examples all run
It would be nice to have a script/action/whatever to run all the examples.  Preferably, this could be run on every PullRequest, or at least, manually run to check for regressions.  Currently, i'm going through all the examples, and many of them don't run without warnings/errors

`fracture_network_solver_scaling/...` often complains about "Package FEHM does not have DelimitedFiles in its dependencies..."
`gw_multigrid_inverse/inputdeck.jl:30` file not found
`piml/theislike_piml_setup.jl:73` params not defined

I haven't checked others

Refs https://github.com/openjournals/joss-reviews/issues/4560
