### 2022-07-12 OrchardLANL/DPFEHM.jl#13 @WilkAndy
`pressure_management/ex.jl`   ERROR: LoadError: UndefVarError: flatten not defined at line 82

`richards/inputdeck.jl`  ERROR: LoadError: BoundsError: attempt to access 640800×640800 SparseArrays.SparseMatrixCSC{Float64, Int64} with 3200798 stored entries at index [1:640800, 640801]

`richards_steadystate_3d/inputdeck.jl`  Similar BoundsError

`scott_example/ex.jl` might be running - i killed it after about 20mins of no output.

`theislike/ex.jl`  ERROR: LoadError: DimensionMismatch("arrays could not be broadcast to a common size; got a dimension with lengths 2401 and 2601") at line 75

I believe all the other examples are working.
