### 2022-07-13 #4560 @omalled
Regarding Wu Hao and Dylan Harp, I got the author list by including the people who made a commit. If I do `git log | grep Ha`, I see Dylan Harp and Wu Hao on there. I haven't checked the code coverage.
