### 2022-07-13 OrchardLANL/DPFEHM.jl#11 @omalled
I think the `Manifest.toml` would get stale if there was one for all the examples. But, if there was one for each example, it could be updated whenever the example is updated. Of course, it would be a pain to retroactively create these `Manifest.toml` files, e.g., in the case you mentioned involving `Preconditioners`. Maybe it's worth it to try though.

The packages [`Project.toml`](https://github.com/OrchardLANL/DPFEHM.jl/blob/master/Project.toml) specifies the julia packages that are needed for the core `DPFEHM` functionality. They are automatically installed with `Pkg.add("DPFEHM")`. However, some of the examples require more packages than that. It looks like `Preconditioners` 0.5.0 is compatible with `AlgebraicMultigrid` 0.4.2. `AlgebraicMultigrid` had some breaking changes (`AlgebraicMultigrid.solve` was renamed to `AlgebraicMultigrid._solve`), so the current version of `DPFEHM` doesn't work with `AlgebraicMultigrid` 0.4.2.

I was able to get [this example](https://github.com/OrchardLANL/DPFEHM.jl/blob/master/examples/fracture_network_solver_scaling/getresults.jl) running by doing the following (from a clean directory with all the `.jl` files in that example):
```
]activate .
add AlgebraicMultigrid Preconditioners
include("getresults.jl")
```
This seems to let the package manager find compatible versions of `AlgebraicMultigrid` and `Preconditioners` without worrying about `DPFEHM`'s requirements. Then I guess it is importing the `DPFEHM` that is installed in the main environment?
