### 2022-07-13 2a7d2fe
fix and comments for gw_multigrid_inverse example
