### 2022-07-13 318e99d
remove old examples that evolved into pressure_management example
