### 2022-07-13 1356435
fix and comments for richards_steadystate_3d, better tests
