### 2022-07-13 9f259c1
comments and simplify richards_anisotropy example
