### 2022-07-13 OrchardLANL/DPFEHM.jl#13 @omalled
OK, I went through those examples and either got it running or deleted the example (some of there were basically deprecated by newer examples). So, I'll close the issue.

There are still some warnings, which I'm OK with. I can try to fix the warnings if you insist, but, e.g., fixing the FEHM warning requires a fix in a different package. I really appreciate you going through these examples! Those Richards examples that weren't working highlighted a testing gap, which I also fixed [here](https://github.com/OrchardLANL/DPFEHM.jl/commit/13564351cdf31822f48dc29f79769ada5306c1dc).
