### 2022-08-21 efbb1e1
add more comments to the examples
