### 2022-08-21 13bca20
add a wave equation test
