+++ paper.md
@@ -57,9 +57,11 @@ They cannot because they do not support automatic differentation, which is neede
 An automatically-differentiable model like `DPFEHM` can be seamlessly integrated into these machine learning workflows.
 This enables machine learning workflows with `DPFEHM` in the loop, e.g., to learn to manage pressure in a scenario where wastewater or carbon dioxide are being injected into the subsurface [@pachalieva2022physics].
 For example, without automatic differentiation the machine learning would get stuck when it needed to compute a Jacobian-vector product involving the subsurface simulator.
-DPFEHM fills this gap and enables the efficient computation of the Jacobian-vector product.
-It is additionally useful for non-machine learning workflows, because gradient calculations are also ubiqitous in more traditional workflows such as inverse analysis [@wu2022inverse] and uncertainty quantification [@betancourt2017conceptual].
+`DPFEHM` fills this gap and enables the efficient computation of the Jacobian-vector product.
+It is additionally useful for non-machine learning workflows, because gradient calculations are also ubiqitous in more traditional workflows such as inverse analysis [@wu2022inverse] and uncertainty quantification [@betancourt2017conceptual] (UQ).
 For example, inverse analysis often uses the gradient to perform some variation of gradient descent to find the solution to the inverse problem, so making this fast is important in this context.
+Traditional inverse modeling and UQ tools (`PEST` [@doherty2010approaches] being the most widely used example), take a non-intrusive approach, which allows them to work with any simulator but forces them to treat the simulator as a black box.
+`DPFEHM` lays the groundwork for next-generation UQ tools that utilize the gradient and Jacobian information that `DPFEHM` efficiently provides.
 Of course, it can also be used to efficiently simulate complex physics related to flow and transport in the subsurface [@greer2022comparison] without exploiting the differentiability very deeply.
 
 An alternative to a differentiable numerical model is to use a differentiable machine learning model that is trained on data from a non-differentiable numerical model such as those listed above.
