+++ paper.bib
@@ -1,3 +1,10 @@
+@book{doherty2010approaches,
+  title={Approaches to highly parameterized inversion: a guide to using PEST for groundwater-model calibration},
+  author={Doherty, John E and Hunt, Randall J},
+  volume={2010},
+  year={2010},
+  publisher={US Department of the Interior, US Geological Survey Reston, VA, USA}
+}
 @techreport{mercer2020amanzi,
   title={Amanzi--ATS: Modeling Environmental Systems across Scales [Brief]},
   author={Mercer-Smith, Janet Anne},
