### 2022-08-22 #4560 @omalled
@WilkAndy, I addressed the issues raised [here](https://github.com/openjournals/joss-reviews/issues/4544#issuecomment-1178305351), following the outline [here](https://github.com/openjournals/joss-reviews/issues/4544#issuecomment-1178365752). I also looked into all the issues raised in the repo and got the code coverage up over 85% (seems to not count code that is run at compile time, so the actual coverage is significantly higher, I believe). Got it done "next week" with a couple of hours to spare in my time zone. 😄 Thanks for looking into all this stuff - definitely helping me whip this code into shape!
