### 2022-10-02 OrchardLANL/DPFEHM.jl#16 @rezgarshakeri
Documentation
In the current documentation (the README file), it is stated that "DPFEHM supports the groundwater flow equations (single phase flow), Richards equation (air/water), the advection-dispersion equation, and the 2d wave equation".
- I think it would be helpful to add those equations to the documentation. Only Richards equation is linked to the [wikipedia](https://en.wikipedia.org/wiki/Richards_equation)  which doesn't match the description. For example, in README you talk about relative permeability and storage, but in that link they are not defined.
- It is mentioned in the paper "DPFEHM uses a two-point flux approximation finite volume scheme". It would be also helpful to add this to the documentation and perhaps its discretized formulation.
- It would be good to add purpose, input, output and variables that are needed inside each function in the DPFEHM.jl/src files. And adding some inline comments would help the user to contribute  in the future.

Refs https://github.com/openjournals/joss-reviews/issues/4560
