### 2022-11-08 OrchardLANL/DPFEHM.jl#16 @omalled
I updated the README to address the first couple bullets: putting the Richards equation in better context with respect to the Wikipedia article and describing the two point flux a bit.

I agree it would be good to add more comments in the source. It could be tricky to do it since many of the functions are generated using metaprogramming. Let me think a bit if there's a good way to deal with that.
