### 2022-12-01 36d8f02
add documentation for relative permeability
