### 2022-12-01 3bd99d0
add documentation to groundwater code
