+++ src/groundwater.jl
@@ -20,18 +20,41 @@
 	end
 end
 
+"""
+amg_solver(A, b; kwargs...)
+
+Return the solution to `A`x=`b` using an algebraic multigrid solver, which the `kwargs` being passed to the AlgebraicMultigrid solver
+"""
 function amg_solver(A, b; kwargs...)
 	ml = AlgebraicMultigrid.ruge_stuben(A)
 	hfree = AlgebraicMultigrid._solve(ml, b; kwargs...)
 	return hfree
 end
 
+"""
+cholesky_solver(A, b; kwargs...)
+
+Return the solution to `A`x=`b` using a Cholesky factorization solver
+"""
 function cholesky_solver(A, b; kwargs...)
 	Af = LinearAlgebra.cholesky(A)
 	hfree = Af \ b
 	return hfree
 end
 
+"""
+groundwater_steadystate(Ks, neighbors, areasoverlengths, dirichletnodes, dirichleths, Qs; linear_solver::Function=amg_solver, kwargs...)
+
+Return the solution to a steady state groundwater flow problem
+
+# Arguments
+- `Ks`: permeability
+- `neighbors`: array of pairs indicating which cells share an interface
+- `areasoverlengths`: array with the same length as `neighbors` that gives the interfacial area divided by the length between the two cell centers
+- `dirichletnodes`: array of indices indicating which nodes have Dirichlet boundary conditions
+- `dirichleths`: array of pressures at the Dirichlet boundary (length is equal to the number of cells on the grid)
+- `Qs`: array of fluxes (length is equal to the number of cells on the grid)
+"""
 function groundwater_steadystate(Ks, neighbors, areasoverlengths, dirichletnodes, dirichleths, Qs; linear_solver::Function=amg_solver, kwargs...)
 	isfreenode, nodei2freenodei, freenodei2nodei = getfreenodes(length(Qs), dirichletnodes)
 	args = (zeros(sum(isfreenode)), Ks, neighbors, areasoverlengths, dirichletnodes, dirichleths, Qs, ones(length(Qs)), ones(length(Qs)))
