### 2022-12-01 a3be77e
add documentation to richards code
