### 2022-12-14 OrchardLANL/DPFEHM.jl#16 @omalled
The code has now been updated with much better documentation throughout. In combination with the previous changes, I think this addresses each of the three bullets. So I'm closing this issue.
