### 2023-06-17 OrchardLANL/DPFEHM.jl#17 @omalled
Thanks, @rtmills! I appreciate your feedback and help in improving the paper.

Item 1: You are correct that the statement was too strong, and I have softened the text. Instead of saying it "cannot" be done, the revised version says it is challenging for high-dimensional problems where one must implement Jacobians to enable reverse-mode (adjoint-based) differentiation. Based on my understanding of the AdjointNet paper (I have discussed these topics with Satish extensively), it utilizes the (non-adjoint) perturbative approach described at the end of page 4 and the beginning of page 5. All the problems they solve have 1 or 2 parameters, so the adjoint method would provide little or no advantage over the perturbative approach in that context.

Item 2: This is a good point; I clarified this at the end of the second paragraph of the statement of need.

Item 3: Thank you for pointing this out. I am not familiar with all these methods, but it is good to mention more possibilities. The text has been modified to reflect this.

Item 4: You are correct that this was overstated. It has been modified to say that "for models that lack discrete adjoints, which covers most non-differentiable models, the cost is...".

Thanks again for your help!
