### 2023-07-04 #4560 @rtmills
@jedbrown and @omalled, thanks for your patience. The changes to the PDF address my concerns, and I am happy to mark this as essentially ready for publication! I will, however, note that there needs to be some editing (that I don't think needs my further review) to one of the changes to the paper , because it currently reads

"[...] Alternative codes such as Amanzi-ATS (Mercer-Smith, 2020), which use a more advanced mimetic finite difference discretization that supports non-K-orthogonal. Other methods for non-K-orthogonal grids include multipoint flux approximations [...]"

and part of the first sentence is clearly missing.

This is very nice work. Thanks for sharing it with the world in JOSS!
