### 2023-07-12 #4560 @jedbrown
Thanks @omalled and sorry to be slow getting to this. This is good work and I just have a few questions.

1. I see that the paper mentions "advection-dispersion" and the repository does too, but [the code](https://github.com/OrchardLANL/DPFEHM.jl/blob/master/src/transport.jl) purports to solve $u_t + \nabla\cdot(\symbf v  u) - \nabla\cdot(D \nabla u) - Q = 0$, which is an advection-diffusion equation. I think this may be a terminology thing in which $D$ is defined in a solution-dependent way to yield dispersion (different wave numbers travel at different speeds), or just that "dispersion" has been overloaded to mean a different thing. I wanted to check that this use is deliberate.

2. The paper discusses applying a Jacobian-vector product, but the adjoint of an implicit method requires a *solve* with the transpose of the Jacobian. Are you using a custom rule to avoid differentiating through the Newton solve and instead only use the transpose-Jacobian at the converged state? Does your implementation rely in assembled matrices?

3. This uses a FV method, for which the discrete adjoint (as produced by AD) is not a consistent discretization of the continuous adjoint. For example, naive upwind 1D advection on a non-uniform periodic 1D grid has zero row sums, but not zero column sums. While this may be useful to optimize a functional of interest, it tends to suffer from spurious local minima and can lead to non-physical solutions. Production FV codes with adjoints (e.g., FUN3D and SU2 for CFD) discretize the continuous adjoint or make related modifications to avoid these serious problems with the discrete adjoint. Have you experienced similar issues and would the framework be able to handle such techniques?

4. It looks like you still have missing DOIs. Can you please add a DOI to all references that have one? [These may help.](https://github.com/openjournals/joss-reviews/issues/4560#issuecomment-1179388331).
