+++ paper.bib
@@ -3,14 +3,16 @@
   author={Doherty, John E and Hunt, Randall J},
   volume={2010},
   year={2010},
-  publisher={US Department of the Interior, US Geological Survey Reston, VA, USA}
+  publisher={US Department of the Interior, US Geological Survey Reston, VA, USA},
+  doi={10.3133/sir20105169}
 }
 @techreport{mercer2020amanzi,
   title={Amanzi--ATS: Modeling Environmental Systems across Scales [Brief]},
   author={Mercer-Smith, Janet Anne},
   number={LA-UR-20-26636},
   year={2020},
-  institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States)}
+  institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States)},
+  doi={10.2172/1657092}
 }
 @article{greer2022comparison,
   title={A Comparison of Linear Solvers for Resolving Flow in Three-Dimensional Discrete Fracture Networks},
@@ -20,13 +22,19 @@
   number={4},
   pages={e2021WR031188},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2021wr031188}
 }
-@article{betancourt2017conceptual,
-  title={A conceptual introduction to Hamiltonian Monte Carlo},
-  author={Betancourt, Michael},
-  journal={arXiv preprint arXiv:1701.02434},
-  year={2017}
+@article{gelman2015stan,
+  title={Stan: A probabilistic programming language for Bayesian inference and optimization},
+  author={Gelman, Andrew and Lee, Daniel and Guo, Jiqiang},
+  journal={Journal of Educational and Behavioral Statistics},
+  volume={40},
+  number={5},
+  pages={530--543},
+  year={2015},
+  publisher={Sage Publications Sage CA: Los Angeles, CA},
+  doi={10.3102/10769986156061}
 }
 @article{wu2022inverse,
   title={Inverse analysis with variational autoencoders: a comparison of shallow and deep networks},
@@ -35,31 +43,39 @@
   volume={3},
   number={2},
   year={2022},
-  publisher={Begel House Inc.}
+  publisher={Begel House Inc.},
+  doi={10.1615/jmachlearnmodelcomput.2022042093}
 }
 @article{pachalieva2022physics,
   title={Physics-informed machine learning with differentiable programming for heterogeneous underground reservoir pressure management},
-  author={Pachalieva, Aleksandra and O'Malley, Daniel and Harp, Dylan Robert and Viswanathan, Hari},
-  journal={arXiv e-prints},
-  pages={arXiv--2206},
-  year={2022}
+  author={Pachalieva, Aleksandra and O’Malley, Daniel and Harp, Dylan Robert and Viswanathan, Hari},
+  journal={Scientific Reports},
+  volume={12},
+  number={1},
+  pages={18734},
+  year={2022},
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41598-022-22832-7}
 }
 @techreport{lichtner2015pflotran,
   title={PFLOTRAN user manual: A massively parallel reactive flow and transport model for describing surface and subsurface processes},
   author={Lichtner, Peter C and Hammond, Glenn E and Lu, Chuan and Karra, Satish and Bisht, Gautam and Andre, Benjamin and Mills, Richard and Kumar, Jitendra},
   year={2015},
-  institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States); Sandia~…}
+  institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States); Sandia~…},
+  doi={10.2172/1168703}
 }
 @techreport{zyvoloski1997summary,
[…]
