### 2023-07-21 #4560 @omalled
1. I am using the word "dispersion" in the same sense as [these notes](https://www3.nd.edu/~bolster/Diogo_Bolster/Ground_Water_files/5-%20ADE.pptx), which is typically how it is used in subsurface flow and transport. We use diffusion to refer to the spreading of mass due to velocity fluctuations at the molecular scale. Disperison, by contrast, is used to describe spreading due to velocity fluctuations at the pore scale (or perhaps a larger scale) that are unresolved by the model. In short, the use is deliberate.

2. Correct, we are using a custom rule to avoid differentiating through the Newton solve and use only the transpose-Jacobian at the converged state. Yes, it relies on assembled Jacobian matrices that are automatically generated.

3. I have been keeping an eye out for issues like this but haven't actually encountered them. That being said, the stuff I've done with the advection-dispersion equation has all been on a uniform, non-periodic grid. It would be possible to use the basic techniques underlying this code to build a continuous adjoint for the backward pass, if needed.

4. I think I fixed all the DOIs this time.

Thanks, @jedbrown!
