### 2023-07-25 #4560 @jedbrown
Looks like you're missing digits in this reference: https://doi.org/10.3102/1076998615606113

Should also protect capitalization for `{Bayesian}`.
