+++ paper.bib
@@ -26,7 +26,7 @@
   doi={10.1029/2021wr031188}
 }
 @article{gelman2015stan,
-  title={Stan: A probabilistic programming language for Bayesian inference and optimization},
+  title={Stan: A probabilistic programming language for {Bayesian} inference and optimization},
   author={Gelman, Andrew and Lee, Daniel and Guo, Jiqiang},
   journal={Journal of Educational and Behavioral Statistics},
   volume={40},
@@ -34,7 +34,7 @@
   pages={530--543},
   year={2015},
   publisher={Sage Publications Sage CA: Los Angeles, CA},
-  doi={10.3102/10769986156061}
+  doi={10.3102/1076998615606113}
 }
 @article{wu2022inverse,
   title={Inverse analysis with variational autoencoders: a comparison of shallow and deep networks},
