### 2023-09-08 #4560 @omalled
I just archived the software on Zenodo: https://zenodo.org/record/8329952

The doi is 10.5281/zenodo.8329952
