+++ examples/fracture_networks_for_qc/decouple_fractures_grid_heterogeneous.jl
@@ -1,0 +1,116 @@
+import DelimitedFiles
+import DPFEHM
+import PyPlot
+
+function addfractures!(logks, fracture_logk, fracture_scales; beta=0.5)
+	#this recursively defines a formula where fracture_logk ∝ length ^ beta as in equation 5 from Jeffrey's paper (10.1002/2016WR018806)
+	if fracture_scales > 0
+		if fracture_logk < 0.0
+			@show fracture_logk
+			@show fracture_scales
+			error("you need to increase fracture_logk so that the permeability of the small fractures is greater than the permeability of the matrix")
+		end
+		logks[div(end, 2), 1:div(3 * end, 4) + 1] .= fracture_logk + log(1.5 ^ beta)#this fracture is 1.5 times longer than the next fracture
+		logks[div(end, 4):div(3 * end, 4), div(end, 2)] .= fracture_logk
+		n = size(logks, 1)
+		#addfractures!(view(logks, 1:div(n, 2), 1:div(n, 2)))#upper left
+		addfractures!(view(logks, 1:div(n, 2), div(n, 2) + 1:n), fracture_logk - log(2 ^ beta), fracture_scales - 1; beta=beta)#upper right
+		#addfractures!(view(logks, div(n, 2) + 1:n, 1:div(n, 2)))#lower left
+		addfractures!(view(logks, div(n, 2) + 1:n, div(n, 2) + 1:n), fracture_logk - log(2 ^ beta), fracture_scales - 1; beta=beta)#lower right
+	end
+end
+addboundaries(logks) = hcat(logks[:, 1], logks, logks[:, end])
+
+function fractal_fractures(N, fracture_scales; doplot=false, matrix_logk=0.0, fracture_logk=1.0, dirichlet=true, neumann=false, caging=false, num_random_nodes=1, kwargs...)
+	mins = [0.0, 0.0]
+	maxs = [1, 1]
+	ns = [N + 2, N]
+	xs = range(mins[1], maxs[1]; length=ns[1])
+	ys = range(mins[2], maxs[2]; length=ns[2])
+	logks = fill(matrix_logk, N, N)
+	addfractures!(logks, fracture_logk, fracture_scales; kwargs...)
+	logks = addboundaries(logks)
+	#make the permeability at the two leftmost columns be the matrix -- this makes it so the dirichlet boundary condition corresponds to the simple Hadamard thing
+	logks[:, 1] .= matrix_logk
+	logks[:, 2] .= matrix_logk
+	neumann_node = 2 * N + div(N, 2)#inject into the tip of the fracture
+	caging_nodes = [div(3 * N, 4) * N + div(N, 4), div(3 * N, 4) * N + div(3 * N, 4)]
+	coords, neighbors, areasoverlengths, _ = DPFEHM.regulargrid2d(mins, maxs, ns, 1.0)
+	random_nodes = rand(1:size(coords, 2), num_random_nodes)
+	if doplot
+		fig, ax = PyPlot.subplots()
+		img = ax.imshow(logks, extent=[mins[1], maxs[1], mins[2], maxs[2]], origin="lower")
+		fig.colorbar(img)
+		if neumann
+			@show coords[:, neumann_node]
+			ax.scatter([coords[1, neumann_node]], [coords[2, neumann_node]], c="r", s=200, alpha=1)
+		elseif caging
+			@show coords[:, neumann_node]
[…]
