### 2023-09-27 9ca9cfd
update fracture network code for qc
